Feature/transfers (#10)
* Transfers impl, as usual a lot of refactoring

* Minor tweaks

+++ app/src/main/java/com/zedorff/yobabooker/ui/views/PieChartView.kt
@@ -390,7 +390,7 @@ class PieChartView(context: Context?, attrs: AttributeSet?) : View(context, attr
                             selectionAnimatorSet.start()
                         }
                         listener?.let {
-                            it.onSliceClick(categories[index].id)
+                            it.onSliceClick(categories[index].id.toInt())
                         }
                     }
                     break
